Refactor streaming_search_inner to return a Stream object
This commit refactors the `streaming_search_inner` method in the `LdapClient` struct to return a `Stream` object instead of a vector of `SearchEntry` objects. The `Stream` object encapsulates the search stream and provides methods for iterating over the search results. This change improves the usability and flexibility of the `streaming_search` method.